pyproject.toml
[tool.poetry]
name = "cacau"
version = "0.1.0"
description = ""
authors = ["marcelo <molivramento@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
django-ninja = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
